starting to fix tests
